requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
requests==2.31.0